+++ test/hexapdf/task/test_pdfa.rb
@@ -38,4 +38,49 @@ describe HexaPDF::Task::PDFA do
     assert_equal('sRGB2014.icc', oi[:Info])
     assert_kind_of(HexaPDF::Stream, oi[:DestOutputProfile])
   end
+
+  describe "fix_glyph_widths" do
+    before do
+      @file = File.join(TEST_DATA_DIR, 'pdfa', 'mismatching_glyph_widths_cidfont_type2.pdf')
+    end
+
+    it "fixes glyph width inconsistencies between the font and the font dictionary" do
+      doc = HexaPDF::Document.open(@file)
+      doc.task(:pdfa, level: '3b', fixes: [:fix_glyph_widths])
+
+      font = HexaPDF::Font::TrueType::Font.new(StringIO.new(doc.object(10).stream))
+      assert_equal(346, font[:hmtx][1].advance_width)
+      doc.dispatch_message(:complete_objects)
+      font = HexaPDF::Font::TrueType::Font.new(StringIO.new(doc.object(10).stream))
+      assert_equal(348, font[:hmtx][1].advance_width)
+    end
+
+    it "works if there is an explicit CIDToGIDMap stream" do
+      doc = HexaPDF::Document.open(@file)
+      doc.object(5)[:CIDToGIDMap] = doc.wrap({}, stream: [0, 1, 2, 3, 4].pack('n*'))
+      doc.task(:pdfa, level: '3b', fixes: [:fix_glyph_widths])
+      doc.dispatch_message(:complete_objects)
+      font = HexaPDF::Font::TrueType::Font.new(StringIO.new(doc.object(10).stream))
+      assert_equal(348, font[:hmtx][1].advance_width)
+    end
+
+    it "processes annotation appearances" do
+      doc = HexaPDF::Document.new
+      doc.pages.add
+      doc.annotations.create_rectangle(doc.pages[0], 20, 20, 20, 60).
+        regenerate_appearance
+      form = doc.pages[0][:Annots][0].create_appearance
+      form.canvas.
+        font(File.join(TEST_DATA_DIR, 'fonts', 'Ubuntu-Title.ttf'), size: 10).
+        text('Hola', at: [0, 0])
+
+      doc = HexaPDF::Document.new(io: StringIO.new(doc.write_to_string))
+      font = doc.pages[0][:Annots][0].appearance.resources.font(:F1).descendant_font
+      font[:W][1][0] = 10
+      doc.task(:pdfa, level: '3b', fixes: [:fix_glyph_widths])
+      doc.dispatch_message(:complete_objects)
+      font = HexaPDF::Font::TrueType::Font.new(StringIO.new(font[:FontDescriptor][:FontFile2].stream))
+      assert_equal(10, font[:hmtx][1].advance_width)
+    end
+  end
 end